Split project in different files
